+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file[^1].
 ## [2.41.0] - 2022-01-14
 ### Added
 - filter by gender to authorities
+- authority matching of swapped names
 
 ### Changed
 - inscription length in item_translations table